Update ExprUtil.kt
Sum and Binary Expression to use prefer Kotlin types.

Update ExpressionTest.kt

Change from Double to Long as Integer is expected.
It shouldn't matter if an INTEGER column is nullable - it is an INTEGER not REAL type.

+++ sqldelight-compiler/src/test/kotlin/app/cash/sqldelight/core/queries/ExpressionTest.kt
@@ -185,7 +185,7 @@ class ExpressionTest {
     )
 
     val query = file.namedQueries.first()
-    assertThat(query.resultColumns.single().javaType).isEqualTo(DOUBLE.copy(nullable = true))
+    assertThat(query.resultColumns.single().javaType).isEqualTo(LONG.copy(nullable = true))
   }
 
   @Test fun `string functions return nullable string only if parameter is nullable`() {
